PBLEID-12100: Store tall - komplettert støtten (ref. review) og litt testdekning

+++ statistics-query-elasticsearch/src/main/java/no/difi/statistics/query/elasticsearch/ElasticsearchQueryService.java
@@ -251,7 +251,7 @@ public class ElasticsearchQueryService implements QueryService {
     private List<Measurement> measurements(SearchHit hit) {
         List<Measurement> measurements = new ArrayList<>();
         hit.getSource().keySet().stream().filter(field -> !field.equals(timeFieldName)).forEach(field -> {
-            int value = Integer.valueOf(hit.getSource().get(field).toString());
+            long value = Long.valueOf(hit.getSource().get(field).toString());
             measurements.add(new Measurement(field, value));
         });
         return measurements;
@@ -260,7 +260,7 @@ public class ElasticsearchQueryService implements QueryService {
     private List<Measurement> measurements(Aggregations sumAggregations) {
         List<Measurement> measurements = new ArrayList<>();
         for (Aggregation sum : sumAggregations) {
-            measurements.add(new Measurement(sum.getName(), (int)((Sum)sum).getValue()));
+            measurements.add(new Measurement(sum.getName(), (long)((Sum)sum).getValue()));
         }
         return measurements;
     }
